requirements.txt
async-timeout==4.0.3
bleak==0.22.2
coloured==1.0.0
dbus-fast==2.24.0
exif==1.6.0
geographiclib==2.0
geopy==2.4.1
goprocam==4.2.0
pillow==10.4.0
plum-py==0.8.7
prettytable==3.11.0
termcolor==2.4.0
typing_extensions==4.12.2
wcwidth==0.2.13
